Support s_curve

+++ Marlin/src/module/calc.h
@@ -23,8 +23,8 @@ void recalc_delta_settings(float radius, float diagonal_rod);
 #ifdef SCARA
 extern float axis_scaling[3];  // Build size scaling
 #endif
-extern float min_pos[3];
-extern float max_pos[3];
+extern double min_pos[3];
+extern double max_pos[3];
 extern float zprobe_zoffset;
 extern int fanSpeed;
 #ifdef BARICUDA
